6.3 작업 내역
- Card 스크립터블 수정 및 리스트 생성 방식 변경
- 카드 Drag&Drop을 통한 카드 사용 + 플레이어, 적 스테이터스 변동
- 적 체력 변동 값에 따른 체력바 UI 제작

+++ Assets/Scripts/Scriptable/Scripts/EntitySO.cs
@@ -8,11 +8,11 @@ public class Status {
     // 이름
     public Sprite main;
     // 이미지
-    public float attack;
+    public int attack;
     // 공격력
-    public float armor;
+    public int armor;
     // 방어도
-    public float health;
+    public int health;
     // 체력
 }
